refactor: SeatSelectionPresenter null -> lateinit
