+++ app/src/main/java/woowacourse/movie/view/reservation/seat/SeatSelectionActivity.kt
@@ -97,7 +97,7 @@ class SeatSelectionActivity :
         binding.btnSeatSelectConfirm.isEnabled = enabled
     }
 
-    override fun showError(message: String) {
+    override fun showError(message: String?) {
         Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
     }
